+++ src/lib/async/SpinLock.h
@@ -12,10 +12,12 @@
 #ifndef SpinLock_include__
 #define SpinLock_include__
 
+#include <cstdint>
+
 class SpinLock {
 private:
-    unsigned long lock;
-    unsigned long* ptr;
+    uint32_t lock;
+    uint32_t* ptr;
 
 public:
     SpinLock(const SpinLock& copy) = delete;  // Verhindere Kopieren
